mica client parsing strategy modification:
* use conf file in rootfs to configure instead of
labels in dockerfile.
* update related container config parser

+++ core/taskService/create.go
@@ -170,7 +170,7 @@ func CreateMicaConf(container *cntr.Container) libmica.MicaClientConf {
 	info := container.GetMicaContainerInfo()
 
 	firmware := info.FirmwarePath()
-	cpu := info.CPU()
+	cpu := uint32(info.CPU())
 	pedestal := info.Ped()
 	name := container.ID
